test(ai): consolidate GP-blocker/peace-target branch pins into matrix (Refs #3749)
Replace the two structurally-parallel observer-phase function-unit branch
suites with one table-driven matrix file, continuing the #3749 branch-pin
consolidation:

- observer_goal_phase_colonial_peace_blocker_branches_test.dart
  (primaryColonialGpBlocker + colonialPhaseGpPeaceTargets, COLONIAL / NW)
- observer_goal_phase_expand_peace_blocker_branches_test.dart
  (primaryInvadableOldWorldGpBlocker + expandPhaseGpPeaceTargets,
  EXPAND / OW)

All four functions share the
({required Game game, required AIWorldSnapshot snapshot}) signature
(String? blocker / List<String> peace targets), so the two blocker
contracts collapse into one truth-table runner and the two peace-target
guard ladders into one case runner. Coverage is preserved 1:1: every
former test becomes one matrix row with the same fixture and verbatim
regression reason (31 rows). The EXPAND-unique minor-first and
mutual-plateau arms are kept as explicit rows; per-family determinism
guards are retained verbatim.

No production code changes; deterministic planning semantics unchanged.

Refs #3749
